+++ tl/src/main/kotlin/com/github/badoualy/telegram/tl/api/TLChannelAdminLogEventActionToggleGroupCallSetting.kt
@@ -2,9 +2,19 @@ package com.github.badoualy.telegram.tl.api
 
 import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_BOOLEAN
 import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_CONSTRUCTOR_ID
+import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_DOUBLE
+import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_INT32
+import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_INT64
+import com.github.badoualy.telegram.tl.TLObjectUtils.computeTLBytesSerializedSize
+import com.github.badoualy.telegram.tl.TLObjectUtils.computeTLStringSerializedSize
 import com.github.badoualy.telegram.tl.serialization.TLDeserializer
 import com.github.badoualy.telegram.tl.serialization.TLSerializer
 import java.io.IOException
+import kotlin.Any
+import kotlin.Boolean
+import kotlin.Int
+import kotlin.String
+import kotlin.jvm.Throws
 
 /**
  * channelAdminLogEventActionToggleGroupCallSetting#56d6a247
@@ -48,6 +58,6 @@ class TLChannelAdminLogEventActionToggleGroupCallSetting() : TLAbsChannelAdminLo
         return joinMuted == other.joinMuted
     }
     companion object  {
-        const val CONSTRUCTOR_ID: Int = 0x56d6a247
+        const val CONSTRUCTOR_ID: Int = 0x56d6a247.toInt()
     }
 }
