+++ tl/src/main/kotlin/com/github/badoualy/telegram/tl/api/TLChannelParticipantsBanned.kt
@@ -1,10 +1,20 @@
 package com.github.badoualy.telegram.tl.api
 
+import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_BOOLEAN
 import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_CONSTRUCTOR_ID
+import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_DOUBLE
+import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_INT32
+import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_INT64
+import com.github.badoualy.telegram.tl.TLObjectUtils.computeTLBytesSerializedSize
 import com.github.badoualy.telegram.tl.TLObjectUtils.computeTLStringSerializedSize
 import com.github.badoualy.telegram.tl.serialization.TLDeserializer
 import com.github.badoualy.telegram.tl.serialization.TLSerializer
 import java.io.IOException
+import kotlin.Any
+import kotlin.Boolean
+import kotlin.Int
+import kotlin.String
+import kotlin.jvm.Throws
 
 /**
  * channelParticipantsBanned#1427a5e1
@@ -48,6 +58,6 @@ class TLChannelParticipantsBanned() : TLAbsChannelParticipantsFilter() {
         return q == other.q
     }
     companion object  {
-        const val CONSTRUCTOR_ID: Int = 0x1427a5e1
+        const val CONSTRUCTOR_ID: Int = 0x1427a5e1.toInt()
     }
 }
