+++ tl/src/main/kotlin/com/github/badoualy/telegram/tl/api/TLDocumentAttributeImageSize.kt
@@ -1,10 +1,20 @@
 package com.github.badoualy.telegram.tl.api
 
+import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_BOOLEAN
 import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_CONSTRUCTOR_ID
+import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_DOUBLE
 import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_INT32
+import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_INT64
+import com.github.badoualy.telegram.tl.TLObjectUtils.computeTLBytesSerializedSize
+import com.github.badoualy.telegram.tl.TLObjectUtils.computeTLStringSerializedSize
 import com.github.badoualy.telegram.tl.serialization.TLDeserializer
 import com.github.badoualy.telegram.tl.serialization.TLSerializer
 import java.io.IOException
+import kotlin.Any
+import kotlin.Boolean
+import kotlin.Int
+import kotlin.String
+import kotlin.jvm.Throws
 
 /**
  * documentAttributeImageSize#6c37c15c
@@ -55,6 +65,6 @@ class TLDocumentAttributeImageSize() : TLAbsDocumentAttribute() {
                 && h == other.h
     }
     companion object  {
-        const val CONSTRUCTOR_ID: Int = 0x6c37c15c
+        const val CONSTRUCTOR_ID: Int = 0x6c37c15c.toInt()
     }
 }
