+++ tl/src/main/kotlin/com/github/badoualy/telegram/tl/api/TLEmojiKeywordsDifference.kt
@@ -1,13 +1,22 @@
 package com.github.badoualy.telegram.tl.api
 
+import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_BOOLEAN
 import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_CONSTRUCTOR_ID
+import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_DOUBLE
 import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_INT32
+import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_INT64
+import com.github.badoualy.telegram.tl.TLObjectUtils.computeTLBytesSerializedSize
 import com.github.badoualy.telegram.tl.TLObjectUtils.computeTLStringSerializedSize
 import com.github.badoualy.telegram.tl.core.TLObject
 import com.github.badoualy.telegram.tl.core.TLObjectVector
 import com.github.badoualy.telegram.tl.serialization.TLDeserializer
 import com.github.badoualy.telegram.tl.serialization.TLSerializer
 import java.io.IOException
+import kotlin.Any
+import kotlin.Boolean
+import kotlin.Int
+import kotlin.String
+import kotlin.jvm.Throws
 
 /**
  * emojiKeywordsDifference#5cc761bd
@@ -77,6 +86,6 @@ class TLEmojiKeywordsDifference() : TLObject() {
                 && keywords == other.keywords
     }
     companion object  {
-        const val CONSTRUCTOR_ID: Int = 0x5cc761bd
+        const val CONSTRUCTOR_ID: Int = 0x5cc761bd.toInt()
     }
 }
