+++ tl/src/main/kotlin/com/github/badoualy/telegram/tl/api/TLEncryptedMessageService.kt
@@ -1,13 +1,22 @@
 package com.github.badoualy.telegram.tl.api
 
+import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_BOOLEAN
 import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_CONSTRUCTOR_ID
+import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_DOUBLE
 import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_INT32
 import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_INT64
 import com.github.badoualy.telegram.tl.TLObjectUtils.computeTLBytesSerializedSize
+import com.github.badoualy.telegram.tl.TLObjectUtils.computeTLStringSerializedSize
 import com.github.badoualy.telegram.tl.core.TLBytes
 import com.github.badoualy.telegram.tl.serialization.TLDeserializer
 import com.github.badoualy.telegram.tl.serialization.TLSerializer
 import java.io.IOException
+import kotlin.Any
+import kotlin.Boolean
+import kotlin.Int
+import kotlin.Long
+import kotlin.String
+import kotlin.jvm.Throws
 
 /**
  * encryptedMessageService#23734b06
@@ -77,6 +86,6 @@ class TLEncryptedMessageService() : TLAbsEncryptedMessage() {
                 && bytes == other.bytes
     }
     companion object  {
-        const val CONSTRUCTOR_ID: Int = 0x23734b06
+        const val CONSTRUCTOR_ID: Int = 0x23734b06.toInt()
     }
 }
