+++ tl/src/main/kotlin/com/github/badoualy/telegram/tl/api/TLGroupCallDiscarded.kt
@@ -1,11 +1,21 @@
 package com.github.badoualy.telegram.tl.api
 
+import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_BOOLEAN
 import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_CONSTRUCTOR_ID
+import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_DOUBLE
 import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_INT32
 import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_INT64
+import com.github.badoualy.telegram.tl.TLObjectUtils.computeTLBytesSerializedSize
+import com.github.badoualy.telegram.tl.TLObjectUtils.computeTLStringSerializedSize
 import com.github.badoualy.telegram.tl.serialization.TLDeserializer
 import com.github.badoualy.telegram.tl.serialization.TLSerializer
 import java.io.IOException
+import kotlin.Any
+import kotlin.Boolean
+import kotlin.Int
+import kotlin.Long
+import kotlin.String
+import kotlin.jvm.Throws
 
 /**
  * groupCallDiscarded#7780bcb4
@@ -67,6 +77,6 @@ class TLGroupCallDiscarded() : TLAbsGroupCall() {
                 && duration == other.duration
     }
     companion object  {
-        const val CONSTRUCTOR_ID: Int = 0x7780bcb4
+        const val CONSTRUCTOR_ID: Int = 0x7780bcb4.toInt()
     }
 }
