+++ tl/src/main/kotlin/com/github/badoualy/telegram/tl/api/TLInputPaymentCredentialsApplePay.kt
@@ -1,9 +1,20 @@
 package com.github.badoualy.telegram.tl.api
 
+import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_BOOLEAN
 import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_CONSTRUCTOR_ID
+import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_DOUBLE
+import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_INT32
+import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_INT64
+import com.github.badoualy.telegram.tl.TLObjectUtils.computeTLBytesSerializedSize
+import com.github.badoualy.telegram.tl.TLObjectUtils.computeTLStringSerializedSize
 import com.github.badoualy.telegram.tl.serialization.TLDeserializer
 import com.github.badoualy.telegram.tl.serialization.TLSerializer
 import java.io.IOException
+import kotlin.Any
+import kotlin.Boolean
+import kotlin.Int
+import kotlin.String
+import kotlin.jvm.Throws
 
 /**
  * inputPaymentCredentialsApplePay#aa1c39f
@@ -47,6 +58,6 @@ class TLInputPaymentCredentialsApplePay() : TLAbsInputPaymentCredentials() {
         return paymentData == other.paymentData
     }
     companion object  {
-        const val CONSTRUCTOR_ID: Int = 0xaa1c39f
+        const val CONSTRUCTOR_ID: Int = 0xaa1c39f.toInt()
     }
 }
