+++ tl/src/main/kotlin/com/github/badoualy/telegram/tl/api/TLKeyboardButtonUrl.kt
@@ -1,10 +1,20 @@
 package com.github.badoualy.telegram.tl.api
 
+import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_BOOLEAN
 import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_CONSTRUCTOR_ID
+import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_DOUBLE
+import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_INT32
+import com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_INT64
+import com.github.badoualy.telegram.tl.TLObjectUtils.computeTLBytesSerializedSize
 import com.github.badoualy.telegram.tl.TLObjectUtils.computeTLStringSerializedSize
 import com.github.badoualy.telegram.tl.serialization.TLDeserializer
 import com.github.badoualy.telegram.tl.serialization.TLSerializer
 import java.io.IOException
+import kotlin.Any
+import kotlin.Boolean
+import kotlin.Int
+import kotlin.String
+import kotlin.jvm.Throws
 
 /**
  * keyboardButtonUrl#258aff05
@@ -55,6 +65,6 @@ class TLKeyboardButtonUrl() : TLAbsKeyboardButton() {
                 && url == other.url
     }
     companion object  {
-        const val CONSTRUCTOR_ID: Int = 0x258aff05
+        const val CONSTRUCTOR_ID: Int = 0x258aff05.toInt()
     }
 }
